Rename and implement connection.
Rename the file to follow PEP-8 standards.
Implement the ability to connect to the AACT database.
